+++ rust/operator-binary/src/operations/mod.rs
@@ -1,2 +1,3 @@
 pub mod graceful_shutdown;
+pub mod job_state;
 pub mod pdb;
